Allow overriding of optflags
Building on top of [@ankane and @ukd1s
conversation](https://github.com/ankane/fastText-ruby/pull/4/files#r661707312)
I chose to explictely leave the -march blank, while still providing affordances
for `with_config` to pull the optflags.

I'm unclear if this is technically more broadly compatible than setting it to
`-march=armv8-a`, but it seems safer to not set something than to incidentally
set it to something that may be invalid?